+++ packages/orq-rc/src/models/operations/createagentrequest.ts
@@ -2572,7 +2572,6 @@ export type CreateAgentRequestResponseBody = {
    * Current semantic version of the agent manifest.
    */
   version?: string | undefined;
-  versionHash?: string | undefined;
   /**
    * Entity storage path in the format: `project/folder/subfolder/...`
    *
@@ -6412,7 +6411,6 @@ export const CreateAgentRequestResponseBody$inboundSchema: z.ZodType<
   updated: z.string().optional(),
   status: CreateAgentRequestStatus$inboundSchema,
   version: z.string().optional(),
-  version_hash: z.string().optional(),
   path: z.string(),
   memory_stores: z.array(z.string()).optional(),
   team_of_agents: z.array(
@@ -6440,7 +6438,6 @@ export const CreateAgentRequestResponseBody$inboundSchema: z.ZodType<
     "project_id": "projectId",
     "created_by_id": "createdById",
     "updated_by_id": "updatedById",
-    "version_hash": "versionHash",
     "memory_stores": "memoryStores",
     "team_of_agents": "teamOfAgents",
     "knowledge_bases": "knowledgeBases",