+++ packages/orq-rc/RELEASES.md
@@ -8199,3 +8199,13 @@ Based on:
 - [typescript v4.8.0-rc.4] packages/orq-rc
 ### Releases
 - [NPM v4.8.0-rc.4] https://www.npmjs.com/package/@orq-ai/node/v/4.8.0-rc.4 - packages/orq-rc
+
+## 2026-04-11 08:48:22
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.3 (2.879.11) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.8.0-rc.5] packages/orq-rc
+### Releases
+- [NPM v4.8.0-rc.5] https://www.npmjs.com/package/@orq-ai/node/v/4.8.0-rc.5 - packages/orq-rc